hw2/requirements.txt
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
google-api-core==2.11.1
google-auth==2.23.0
google-cloud-core==2.3.3
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
idna==3.4
networkx==3.1
numpy==1.26.0
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
scipy==1.11.2
soupsieve==2.5
tqdm==4.66.1
urllib3==1.26.16
